swal

+++ src/Components/Users.jsx
@@ -2,6 +2,7 @@
 import { Link } from 'react-router-dom';
 import { useQuery } from '@tanstack/react-query';
 import useTitle from '../hooks/useTitle';
+import Swal from 'sweetalert2';
 
 const Users = () => {
     useTitle('Home');
@@ -13,17 +14,34 @@ const Users = () => {
 
     
     const handleDelete = id => {
-        fetch(`http://localhost:5000/users/${id}`, {
-            method: 'DELETE'
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+            }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:5000/users/${id}`, {
+                    method: 'DELETE'
+                })
+                .then(res => res.json())
+                .then(data => {
+                    console.log(data);
+                    if (data.deletedCount > 0) {
+                        Swal.fire(
+                            'Deleted!',
+                            'Information has been deleted.',
+                            'success'
+                        )
+                    }
+                    refetch();
+                })
+            }
         })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data);
-                if (data.deletedCount > 0) {
-                    alert("Delete Successful");
-                }
-                refetch();
-            })
+        
     }
 
     return (
